+++ R/calc_indicator.R
@@ -59,7 +59,6 @@ calc_indicators <- function(x, indicators, ...) {
 #' @keywords internal
 #' @importFrom dplyr relocate last_col
 #' @importFrom tidyr nest
-#' @importFrom data.table rbindlist
 .get_single_indicator <- function(x, indicator, ...) {
   i <- NULL
   # get arguments from function call and portfolio object
@@ -108,7 +107,18 @@ calc_indicators <- function(x, indicators, ...) {
   unlink(file.path(tmpdir, "terra"), recursive = TRUE, force = TRUE)
   terra::terraOptions(tempdir = terra_org)
   # bind results to data.frame
-  results <- tibble(data.table::rbindlist(results, fill = TRUE, idcol = ".id"))
+  index_tbl <- purrr::map_lgl(results, function(x) inherits(x, "tbl_df"))
+  if(any(index_tbl) & any(!index_tbl)){
+    colnames <- names(results[[which(index_tbl)[1]]])
+    fill <- rep(NA, length(colnames))
+    names(fill) <- colnames
+    for (i in which(!index_tbl)) results[[i]] <- fill
+    results <- tibble(dplyr::bind_rows(results, .id = ".id"))
+  } else if (any(!index_tbl)) {
+    results <- tibble(.id = x$assetid, value = rep(NA, nrow(x)))
+  } else {
+    results <- tibble(dplyr::bind_rows(results, .id = ".id"))
+  }
   # nest the results
   results <- nest(results, !!indicator := !.id)
   # attach results